Added connection enum and refined reconnect logic
Even with the PING/PONG twitch will disconnect for unknown reasons leaving a "Protocol(ResetWithoutClosingHandshake)" error message.  This update sets the connection state to disconnected, waits 5 seconds then attempts a new connection.  As the listener uses a new random justinfan account each time there is no concern of a lockout using this technique however with the writer account this will need to be handled more diligently.