Fixed bug in parsing message
After updating the amount of meta data that was being returned on the message, there was an issue where it wasn't correctly splitting the message because the first ":" may or may not have a space before it " :".  This was corrected by removing the space, however it was determined that some of the extra meta information was needed to get the Display name.  Also minor refactoring of variable names.