Conversion for the LSB XML to CSV
This is included for reference only.  This script shows a step by step approach to changing a file format to CSV and handling the various problems that can arise from custom file formats.  Note LSB files are not included as they are copywritten and you will need to request your own permission to use them.  ChapterVerse has obtained that permission and it will be available in the production version.